// js/stat-evaluation-player/src/generated/BallThirdTeamStats.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LabeledFloatSums } from "./LabeledFloatSums.ts";

export type BallThirdTeamStats = { tracked_time: number, defensive_third_time: number, neutral_third_time: number, offensive_third_time: number, labeled_time?: LabeledFloatSums, };